refactor: MatchResult 팩토리 메서드 제거 및 점수 비교 책임을 딜러에게 위임
