+++ src/main/kotlin/blackjack/GameParticipants.kt
@@ -8,15 +8,9 @@ data class GameParticipants(
 
     fun calcMatchResult(): GameParticipantResults {
         val gameParticipantResults = players.map { player ->
-            val matchResult = compareScore(player, dealer)
+            val matchResult = dealer.compareScore(player)
             GameParticipantPlayerResult(player.name, matchResult, player.betAmount)
         }
         return GameParticipantResults(gameParticipantResults)
     }
-
-    private fun compareScore(player: GameParticipantPlayer, dealer: GameParticipantDealer): MatchResult =
-        if (player.isBust || dealer.isBlackjack()) MatchResult.LOSS
-        else if (player.isBlackjack()) MatchResult.BLACKJACK
-        else if (dealer.isBust) MatchResult.WIN
-        else MatchResult.of(player, dealer)
 }
