+++ src/main/kotlin/blackjack/MatchResult.kt
@@ -1,21 +1,12 @@
 package blackjack
 
-typealias CompareScore = (GameParticipantPlayer, GameParticipantDealer) -> Boolean
-
 enum class MatchResult(
     val message: String,
-    val rate: Double,
-    private val compare: CompareScore
+    val rate: Double
 ) {
-    BLACKJACK("승", 1.5, { player, dealer -> player.isBlackjack() }),
-    WIN("승", 1.0, { player, dealer -> player > dealer }),
-    LOSS("패", -1.0, { player, dealer -> player < dealer }),
-    DRAW("무", 0.0, { player, dealer -> player.compareTo(dealer) == 0 })
+    BLACKJACK("승", 1.5),
+    WIN("승", 1.0),
+    LOSS("패", -1.0),
+    DRAW("무", 0.0)
     ;
-
-    companion object {
-        fun of(player: GameParticipantPlayer, dealer: GameParticipantDealer): MatchResult =
-            MatchResult.values().find { it.compare(player, dealer) }
-                ?: throw RuntimeException()
-    }
 }
